requirements.txt
# Core LLM + AI APIs

openai>=1.0.0
elevenlabs==0.2.27



#Frontend and UI
streamlit>=1.27.0

#Resume Parsing
PyMuPDF>-1.22.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0

#Audio I/O
speechrecognition>=3.10.0   
sounddevice>=0.4.6         
soundfile>=0.12.1         


python-dotenv>=1.0.0       
numpy>=1.25.2               

requests       
tqdm